requirements.txt
Django~=4.1.7
Pillow==9.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
djangorestframework~=3.14.0
pytz~=2023.2
django-import-export==2.4.0